Allow content in the body element
Prior to this commit, the only place where content was allowed was a
textarea element.

With this commit, content may be placed directly under the body element.
This should be done with care so that there is no HTML syntax error in
the comment. These caveats are explained in README.