Remove hack for commonmark loading test
The test to check that window.commonmark is defined after commonmark
loads in browser had a hack to workaround an issue due to which jsdom
fired the window.onload event before a script element's onload event
even though the script element was added to the DOM as part of the
window load. The workaround involved waiting for a fixed amount of time
to wait for the script to load before calling assert functions to check
that window.commonmark is defined.

This issue existed in jsdom 11 but does not exist in jsdom 13 now. So
this commit gets rid of this hack and implements this test in the
natural way, i.e., test that window.commonmark is defined after the
window.onload event fires.