Support escaped dollar for literal dollar sign
The MathJax configuration in this project has `processEscapes` set to
`true`. This means that in the text, `\$` can be used to represent a
literal dollar sign. However, prior to this change, this project was not
parsing for these `\$` tokens and treating them differently.

So something like `\$ 1 + 1 = 2 $` could be considered to be a single
token to be kept intact for MathJax. With this fix, `\$` would be
considered a single token to be kept intact for MathJax (which in turn
would treat it as a literal dollar sign) and ` 1 + 1 = 2 $` would be
considered Markdown text.