Replace Node.js v7 with v8 in the build matrix
The new version of `ws` package, i.e., `ws 7.0.0` uses object spread
syntax which is not supported in Node v7. This package is installed as a
dependency of the `jsdom` package which we depend on.

`ws 6.2.1` [did not use the object spread syntax][1], so the Travis CI
builds for this project used to run fine with Node v7.

`ws 7.0.0` [does use the object spread syntax][2], so the Travis CI
builds for this project fails with Node v7 now. The following error
occurs:

    $ npm test

    > texme@0.7.0 test /home/travis/build/susam/texme
    > standard && mocha

    /home/travis/build/susam/texme/node_modules/ws/lib/websocket.js:347
          ...options
          ^^^
    SyntaxError: Unexpected token ...

The object spread syntax is [available since][3] Node.js v8.60 only.
Therefore, resolve this issue by updating the Travis CI build matrix to
use Node.js v8 instead of v7.

[1]: https://github.com/websockets/ws/blob/6.2.1/lib/websocket.js#L347
[2]: https://github.com/websockets/ws/blob/7.0.0/lib/websocket.js#L347
[3]: https://node.green/