Fix tokenization of nested LaTeX environments
The following input file reproduces the issue:

    <!DOCTYPE html><script src="texme.js"></script><textarea>
    \begin{align*}
    A & = \begin{bmatrix}
            1 & 0 \\
            0 & 1
          \end{bmatrix} \\
    {a}_{i} & = {b}_{i}
    \end{align*}

Without this fix, the `align` environment is masked only partially as
follows:

    ::MASK:: \\
    {a}_{i} & = {b}_{i}
    \end{align*}

The masking ends as soon as the first `\end` command is encountered. As
a result, after Markdown rendering and unmasking, the output looks like
this:

    <p>\begin{align*}
    A & = \begin{bmatrix}
            1 & 0 \\
            0 & 1
          \end{bmatrix} \
    {a}<em>{i} &amp; = {b}</em>{i}
    \end{align*}</p>

This is an issue because `{a}_{i} & = {b}_{i}` has been rendered as
Markdown, thus introducing unwanted `<em>` tags. Further, this is a
broken input for MatJax. MathJax succeeds in rendering only the
`bmatrix` environment correctly. It fails to render the `align*`
environment.

With this fix, the entire `align` environment is masked as expected:

    ::MASK::

Thus, after Markdown rendering and unmasking, the output is a valid
input for MathJax:

    <p>\begin{align*}
    A & = \begin{bmatrix}
            1 & 0 \\
            0 & 1
          \end{bmatrix} \\
    {a}_{i} & = {b}_{i}
    \end{align*}</p>

MathJax renders this successfully.